Dataset describing the parameters for the epidemiological report production

A dataset describing the parameters to be used for each output of each disease report for all 53 health topics included in TESSy

Usage

AERparams

Format

A data frame with 53 rows (corresponding to the 53 health topics) and 24 variables:

HealthTopic

Disease code that should match with the health topic code from the disease-specific dataset e.g. ANTH, SALM, etc.

DG

(optional) Disease group e.g. FWD

DP

(optional) Disease programme e.g. FWD

Label

Disease label to be used in the report e.g. salmonellosis, anthrax

FrequencyCategory

(optional) Frequency of the disease e.g. VERY RARE, NON-RARE, etc.

MeasurePopulation

Type of population presented for this disease i.e. ALL or CONFIRMED cases

DatePublicAtlas

Date of latest availability in the public access of the Atlas

TableUse

Type of table to present in the report i.e. NO table, ASR table presenting age-standardised rates, RATE table presenting rates or COUNT table presenting the number of cases only.

TableRatesLabel

Label to use in the table for rates e.g. RATE PER 100000 POPULATION

TableRatesNoDecimals

Number of decimals to use when presenting rates

TableASRNoDecimals

Number of descimals to use when presenting ASR

AgeGenderUse

Type of age and gender bar graph to present i.e. NO graph, AG-COUNT Bar graph presenting the number of cases by age and gender, AG-RATE Bar graph presenting the rates of cases by age and gender, AG-PROP Bar graph presenting the proportion of cases by age and gender, A-RATE Bar graph presenting the rates of cases by age.

AgeGenderBarGraphLabel

Label to use in the age and gender bar graph

AgeGenderGraphNoDecimals

Number of decimals to use when presenting rates in the age and gender bar graph

TSTrendGraphUse

Logical Y/N specifying whether to include a line graph describing the trend of the disease over the time

TSSeasonalityGraphUse

Logical Y/N specifying whether to include a line graph describing the seasonality of the disease

TSSpecific

Logical Y/N for specific line graph inclusion

MapNumbersUse

Logical Y/N specifying whether to include the map presenting the number of cases by Member State

MapRatesUse

Logical Y/N specifying whether to include the map presenting the rates of cases by Member State

MapRatesNoDecimals

(optional) Number of decimals to use for presenting maps

MapASRUse

Logical Y/N specifying whether to include the map presenting the age-standardised rates of cases by Member State

MapASRNoDecimals

(optional) Number of decimals to use for presenting maps

Transmission

Not implemented yet

TransmissionNoDecimals

Not implemented yet