Dataset including Dengue data for 2015-2019

A dataset containing the data and indicators required to build the epidemiological report for Dengue 2019 TESSy data (default dataset used throughout EpiReport)

Usage

DENGUE2019

Format

A data frame with 44,332 rows and 11 variables:

HealthTopicCode

Disease code e.g. ANTH, SALM, etc.

MeasureCode

Code of the measure indicator

TimeUnit

Unit of the time variable i.e. Y for yearly data or M for monthly data

TimeCode

Time variable including dates in any formats available (according to the unit defined in TimeUnit) yearly data (e.g. 2001) or monthly data (e.g. 2001-01)

GeoCode

Geographical level in coded format including country names (e.g. AT for Austria, BE for Belgium, BG for Bulgaria, see also the EpiReport::MSCode table, correspondence table for Member State labels and codes)

XValue

XValue

XLabel

The label associated with the x-axis in the epidemiological report (see getAgeGender() and plotAgeGender() bar graph for the age variable)

YValue

The value associated with the y-axis in the epidemiological report (see plotAge() bar graph for the variable age, or getTableByMS() for the number of cases, rate or age-standardised rate in the table by Member States by year)

YLabel

The label associated with the y-axis in the epidemiological report (see getAgeGender() and plotAgeGender() bar graph for the grouping variable gender)

ZValue

The value associated with the stratification of XLabel and YLabel in the age and gender bar graph (see getAgeGender() and plotAgeGender())

N

Number of cases (see getTrend() and getSeason() line graph)

See also