Dataset correspondence table between country names and country code

Dataframe providing the correspondence table of the geographical code GeoCode used in the disease dataset, and the geographical label Country to use throughout the report. Additional information on the EU/EEA affiliation is also available in column EUEEA.

Usage

MSCode

Format

A data frame with 32 rows and 3 variables:

Country

Full name of the country / Member State e.g. Austria, Belgium, etc.

TheCountry

Full name of the country / Member State including 'the' article for NL and UK e.g. Austria, Belgium, the Netherlands, the United Kingdom etc.

GeoCode

Associated code (see GeoCode variable on the SALM2016 internal dataset) e.g. AT, BE, BG, etc.

EUEEA

For each Member State, variable specifying in the country is part of the EU or EEA.

See also