Get full disease-specific epidemiological report

Function to generate the 'Microsoft Word' epidemiological report (similar to the ECDC Annual Epidemiological Report (AER)) including all disease-specific outputs at each output-specific bookmarks exact location.
(for further information on the outputs and the corresponding bookmarks, please see the package vignette "The Epidemiological Report Package" with browseVignettes("EpiReport"))
(see ECDC AER https://www.ecdc.europa.eu/en/publications-data/monitoring/all-annual-epidemiological-reports)

Usage

getAER(
  template = file.path(system.file(package = "EpiReport"), "template/AER_template.docx"),
  outputPath = getwd(),
  x = EpiReport::DENGUE2019,
  disease = "DENGUE",
  year = 2019,
  reportParameters = EpiReport::AERparams,
  MSCode = EpiReport::MSCode,
  pathPNG = system.file("maps", package = "EpiReport")
)

Arguments

template

doc (see 'officer' package), the empty 'Word' document template in which to include the table and plots disease-specific outputs. Default value is the empty template included in the package. See getTemplate().

outputPath

character string, the full path where to generate the epidemiological report 'Word' output. Default value is the current working directory getwd().

x

dataframe, raw disease-specific dataset (see specification of the dataset in the package vignette with browseVignettes("EpiReport")) (default DENGUE2019)

disease

character string, disease code (default "DENGUE"). Please make sure the disease code is included in the disease-specific dataset x in the HealthTopicCode variable.

year

numeric, year to produce the report for (default 2019). Please make sure the year is included in the disease-specific dataset x in the TimeCode variable.

reportParameters

dataframe, dataset including the required parameters for the report production (default AERparams) (see specification of the dataset in the package vignette with browseVignettes(package = "EpiReport"))

MSCode

dataframe, correspondence table of GeoCode names and codes (default MSCode) (see specification of the dataset in the package vignette with browseVignettes(package = "EpiReport"))

pathPNG

character string, the full path to the folder containing the maps (in PNG) to include in the final report

Value

A 'Word' document

See also

Examples

if (FALSE) { # \dontrun{
# --- Generating the AER report using the default Dengue dataset
getAER()
} # }

if (FALSE) { # \dontrun{
# --- Or using external data (example below)
ZIKV2016 <- read.table("data/ZIKV2016.csv", sep = ",", header = TRUE, stringsAsFactors = FALSE)
output <- "C:/EpiReport/doc/"
pathMap <- "C:/EpiReport/maps/"
getAER(disease = "ZIKV", year = 2016, x = ZIKV2016, outputPath = output, pathPNG = pathMap)
} # }