Get disease-specific table: distribution of cases by Member State (GeoCode)

Function returning the table ('flextable') that will be included in the epidemiological report at the bookmark location 'TABLE1' of the template report. An additional caption will be included at the location of the bookmark 'TABLE1_CAPTION'.
(see Table 1 of the ECDC annual reports https://www.ecdc.europa.eu/en/publications-data/monitoring/all-annual-epidemiological-reports)

Usage

getTableByMS(
  x = EpiReport::DENGUE2019,
  disease = "DENGUE",
  year = 2019,
  reportParameters = EpiReport::AERparams,
  MSCode = EpiReport::MSCode,
  index = 1,
  doc
)

Arguments

x

dataframe, raw disease-specific dataset (see specification of the dataset in the package vignette with browseVignettes(package = "EpiReport")) (default DENGUE2019)

disease

character string, disease code (default "DENGUE"). Please make sure the disease code is included in the disease-specific dataset x in the HealthTopicCode variable.

year

numeric, year to produce the table for (default 2019). Please make sure the year is included in the disease-specific dataset x in the TimeCode variable.

reportParameters

dataframe, dataset including the required parameters for the report production (default AERparams) (see specification of the dataset in the package vignette with browseVignettes(package = "EpiReport"))

MSCode

dataframe, correspondence table of GeoCode names and codes (default MSCode) (see specification of the dataset in the package vignette with browseVignettes(package = "EpiReport"))

index

integer, figure number

doc

'Word' document (see officer package) in which to add the table at the bookmark location. If doc is missing, getTable returns the flextable table object.

Value

'Word' doc or flextable object (see 'flextable' package)

Details

The current version of the 'EpiReport' package includes three types of table (see detailed specification of the tables in the package vignette with browseVignettes(package = "EpiReport")):

COUNT - Table presenting the number of cases by Member State (GeoCode) over a 5-year period;

RATE - Table presenting the number of cases and rates by Member State (GeoCode) over a 5-year period;

ASR - Table presenting the number of cases and rates by Member State (GeoCode) over a 5-year period, including age-standardised rates for the most recent year.

See also

Examples

# --- Draft the table using the default Dengue dataset
getTableByMS()

Country

2015

2016

2017

2018

2019

Number

Rate

Number

Rate

Number

Rate

Number

Rate

Number

Rate

ASR

Austria

 103

1.2

 116

1.3

  85

1.0

  85

1.0

 142

1.6

1.7

Belgium

 108

1.0

 114

1.0

  77

0.7

 101

0.9

 202

1.8

1.9

Bulgaria

.

.

.

.

.

.

.

.

.

.

.

Croatia

-

-

   2

0.0

   0

0.0

   2

0.0

   4

0.1

0.1

Cyprus

.

.

.

.

.

.

.

.

.

.

.

Czechia

   0

0.0

   0

0.0

   0

0.0

   0

0.0

   0

0.0

0.0

Denmark

.

.

.

.

.

.

.

.

.

.

.

Estonia

  12

0.9

   9

0.7

   8

0.6

   6

0.5

   6

0.5

0.5

Finland

  54

1.0

  66

1.2

  25

0.5

  56

1.0

  81

1.5

1.6

France

 285

0.4

 373

0.6

 266

0.4

 331

0.5

 904

1.3

1.4

Germany

 722

0.9

 958

1.2

 635

0.8

 614

0.7

1175

1.4

1.5

Greece

   2

0.0

   2

0.0

   1

0.0

   2

0.0

  10

0.1

0.1

Hungary

  12

0.1

  24

0.2

  17

0.2

  14

0.1

  44

0.5

0.5

Iceland

   0

0.0

   0

0.0

   1

0.3

   1

0.3

   4

1.1

1.1

Ireland

   8

0.2

  18

0.4

  10

0.2

  17

0.4

  18

0.4

0.4

Italy

 103

0.2

 106

0.2

  95

0.2

 108

0.2

 231

0.4

0.4

Latvia

   4

0.2

   9

0.5

  13

0.7

  12

0.6

  11

0.6

0.6

Liechtenstein

.

.

.

.

.

.

.

.

.

.

.

Lithuania

   9

0.3

   4

0.1

   4

0.1

   8

0.3

   9

0.3

0.4

Luxembourg

   0

0.0

   1

0.2

   0

0.0

   1

0.2

   1

0.2

0.2

Malta

   1

0.2

   1

0.2

   3

0.7

   1

0.2

   2

0.4

0.4

Netherlands

  18

-

   6

-

   0

-

   0

-

   0

-

-

Norway

  98

1.9

  64

1.2

  35

0.7

  49

0.9

 102

1.9

2.0

Poland

  12

0.0

  41

0.1

  29

0.1

  30

0.1

  55

0.1

0.1

Portugal

  14

0.1

  13

0.1

  11

0.1

  14

0.1

  30

0.3

0.3

Romania

   7

0.0

   8

0.0

   7

0.0

   4

0.0

  15

0.1

0.1

Slovakia

   2

0.0

   4

0.1

   2

0.0

   7

0.1

   6

0.1

0.1

Slovenia

   3

0.1

   6

0.3

   5

0.2

   8

0.4

  21

1.0

1.1

Spain

 168

0.4

 261

0.6

 128

0.3

 205

0.4

 228

0.5

0.5

Sweden

 159

1.6

 225

2.3

 106

1.1

 106

1.0

 235

2.3

2.4

United Kingdom

 423

0.7

 468

0.7

 465

0.7

 432

0.7

 827

1.2

1.3

EU-EEA

2327

0.5

2899

0.6

2028

0.4

2214

0.5

4363

0.9

0.9